scripts/requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
dnspython==2.6.1
idna==3.7
numpy==1.24.4
pandas==2.0.3
pymongo==4.7.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
tabulate==0.9.0
tzdata==2024.1
urllib3==2.2.1
wbgapi==1.0.12
